+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [1.2.0]
 
 ### Additions
+- [Issue #4] Add method to close winston logger connection
 
 ### Fixes